requirements.txt
flask==3.0.2
boto3==1.34.69
pandas==2.2.1
python-dotenv==1.0.1
gzip-reader==0.1.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
